+++ addons/l10n_at/models/__init__.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import datev
 from . import chart_template
-from . import ir_actions_report
-from . import res_company
