+++ addons/l10n_ch/models/__init__.py
@@ -9,3 +9,4 @@ from . import res_bank
 from . import res_company
 from . import account_bank_statement
 from . import ir_actions_report
+from . import chart_template
