+++ addons/l10n_de/__manifest__.py
@@ -16,19 +16,13 @@ German accounting chart and localization.
         'account',
         'base_iban',
         'base_vat',
+        'l10n_din5008',
     ],
     'data': [
         'data/account_account_tags_data.xml',
         'data/menuitem_data.xml',
         'views/account_view.xml',
         'views/res_company_views.xml',
-        'report/din5008_report.xml',
-        'data/report_layout.xml',
     ],
-    'assets': {
-        'web.report_assets_common': [
-            'l10n_de/static/src/**/*',
-        ],
-    },
     'license': 'LGPL-3',
 }
