--- addons/l10n_de/data/report_layout.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-        <record id="report_layout_din5008" model="report.layout">
-            <field name="view_id" ref="l10n_de.external_layout_din5008"/>
-            <field name="image">/l10n_de/static/img/preview_din.png</field>
-            <field name="pdf">/l10n_de/static/pdf/preview_din.pdf</field>
-            <field name="name">DIN 5008</field>
-        </record>
-    </data>
-</odoo>
