+++ addons/l10n_de/models/chart_template.py
@@ -17,7 +17,12 @@ class AccountChartTemplate(models.Model):
     # Write paperformat and report template used on company
     def _load(self, sale_tax_rate, purchase_tax_rate, company):
         res = super(AccountChartTemplate, self)._load(sale_tax_rate, purchase_tax_rate, company)
-        if company.account_fiscal_country_id.code == 'DE':
-            company.write({'external_report_layout_id': self.env.ref('l10n_de.external_layout_din5008').id,
-            'paperformat_id': self.env.ref('l10n_de.paperformat_euro_din').id})
+        if self in [
+            self.env.ref('l10n_de_skr03.l10n_de_chart_template', raise_if_not_found=False),
+            self.env.ref('l10n_de_skr04.l10n_chart_de_skr04', raise_if_not_found=False)
+        ]:
+            company.write({
+                'external_report_layout_id': self.env.ref('l10n_din5008.external_layout_din5008').id,
+                'paperformat_id': self.env.ref('l10n_din5008.paperformat_euro_din').id
+            })
         return res
