--- addons/l10n_de_purchase/__manifest__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-{
-    'name': 'Germany - Purchase',
-    'category': 'Accounting/Localizations',
-    'depends': [
-        'l10n_de',
-        'purchase',
-    ],
-    'auto_install': True,
-    'license': 'LGPL-3',
-}
