--- addons/l10n_de_purchase/i18n/l10n_de_purchase.pot
@@ -1,114 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* l10n_de_purchase
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-07-22 09:49+0000\n"
-"PO-Revision-Date: 2021-07-22 09:49+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Cancelled Purchase Order"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Cancelled Purchase Order No."
-msgstr ""
-
-#. module: l10n_de_purchase
-#: model:ir.model.fields,field_description:l10n_de_purchase.field_purchase_order__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: model:ir.model.fields,field_description:l10n_de_purchase.field_purchase_order__id
-msgid "ID"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Incoterm"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: model:ir.model.fields,field_description:l10n_de_purchase.field_purchase_order__l10n_de_addresses
-msgid "L10N De Addresses"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: model:ir.model.fields,field_description:l10n_de_purchase.field_purchase_order__l10n_de_document_title
-msgid "L10N De Document Title"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: model:ir.model.fields,field_description:l10n_de_purchase.field_purchase_order__l10n_de_template_data
-msgid "L10N De Template Data"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: model:ir.model.fields,field_description:l10n_de_purchase.field_purchase_order____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Order Date"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Order Reference"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#: model:ir.model,name:l10n_de_purchase.model_purchase_order
-#, python-format
-msgid "Purchase Order"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Purchase Order No."
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Purchase Representative"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Request for Quotation"
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Request for Quotation No."
-msgstr ""
-
-#. module: l10n_de_purchase
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#: code:addons/l10n_de_purchase/models/purchase.py:0
-#, python-format
-msgid "Shipping Address:"
-msgstr ""
