--- addons/l10n_de_repair/__manifest__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-{
-    'name': 'Germany - Repair',
-    'category': 'Accounting/Localizations',
-    'depends': [
-        'l10n_de',
-        'repair',
-    ],
-    'auto_install': True,
-    'license': 'LGPL-3',
-}
