--- addons/l10n_de_sale/__manifest__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-{
-    'name': 'Germany - Sale',
-    'category': 'Accounting/Localizations',
-    'depends': [
-        'l10n_de',
-        'sale',
-    ],
-    'auto_install': True,
-    'license': 'LGPL-3',
-}
