--- addons/l10n_de_sale/i18n/de.po
@@ -1,121 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* l10n_de_sale
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-06-30 07:09+0000\n"
-"PO-Revision-Date: 2021-07-29 13:08+0200\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-"X-Generator: Poedit 2.3\n"
-"Last-Translator: \n"
-"Language: de\n"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Customer Reference"
-msgstr "Kundenreferenz"
-
-#. module: l10n_de_sale
-#: model:ir.model.fields,field_description:l10n_de_sale.field_sale_order__display_name
-msgid "Display Name"
-msgstr "Anzeigename"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Expiration"
-msgstr "Ablaufdatum"
-
-#. module: l10n_de_sale
-#: model:ir.model.fields,field_description:l10n_de_sale.field_sale_order__id
-msgid "ID"
-msgstr "ID"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Invoicing Address:"
-msgstr "Rechnungsadresse:"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Invoicing and Shipping Address:"
-msgstr "Rechnungs- und Lieferadresse:"
-
-#. module: l10n_de_sale
-#: model:ir.model.fields,field_description:l10n_de_sale.field_sale_order__l10n_de_addresses
-msgid "L10N De Addresses"
-msgstr ""
-
-#. module: l10n_de_sale
-#: model:ir.model.fields,field_description:l10n_de_sale.field_sale_order__l10n_de_document_title
-msgid "L10N De Document Title"
-msgstr ""
-
-#. module: l10n_de_sale
-#: model:ir.model.fields,field_description:l10n_de_sale.field_sale_order__l10n_de_template_data
-msgid "L10N De Template Data"
-msgstr ""
-
-#. module: l10n_de_sale
-#: model:ir.model.fields,field_description:l10n_de_sale.field_sale_order____last_update
-msgid "Last Modified on"
-msgstr "Zuletzt bearbeitet am"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Order Date"
-msgstr "Auftragsdatum"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Order No."
-msgstr "Auftragsnummer"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Quotation"
-msgstr "Angebot"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Quotation Date"
-msgstr "Angebotsdatum"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Quotation No."
-msgstr "Angebotsnummer"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#: model:ir.model,name:l10n_de_sale.model_sale_order
-#, python-format
-msgid "Sales Order"
-msgstr "Verkaufsauftrag"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Salesperson"
-msgstr "Verkäufer"
-
-#. module: l10n_de_sale
-#: code:addons/l10n_de_sale/models/sale.py:0
-#, python-format
-msgid "Shipping Address:"
-msgstr "Lieferadresse:"
