--- addons/l10n_de_stock/__manifest__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-{
-    'name': 'Germany - Stock',
-    'category': 'Accounting/Localizations',
-    'depends': [
-        'l10n_de',
-        'stock',
-    ],
-    'auto_install': True,
-    'license': 'LGPL-3',
-}
