--- addons/l10n_de_stock/i18n/de.po
@@ -1,49 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* l10n_de_stock
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-06-30 07:09+0000\n"
-"PO-Revision-Date: 2021-07-15 11:38+0200\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-"X-Generator: Poedit 3.0\n"
-"Last-Translator: \n"
-"Language: de\n"
-
-#. module: l10n_de_stock
-#: code:addons/l10n_de_stock/models/stock.py:0
-#, python-format
-msgid "Date"
-msgstr "Datum"
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory__display_name
-msgid "Display Name"
-msgstr "Anzeigename"
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory__id
-msgid "ID"
-msgstr "ID"
-
-#. module: l10n_de_stock
-#: model:ir.model,name:l10n_de_stock.model_stock_inventory
-msgid "Inventory"
-msgstr "Lager"
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory__l10n_de_template_data
-msgid "L10N De Template Data"
-msgstr "L10N De Template Data"
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory____last_update
-msgid "Last Modified on"
-msgstr "Zuletzt bearbeitet am"
