--- addons/l10n_de_stock/i18n/l10n_de_stock.pot
@@ -1,47 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* l10n_de_stock
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-07-22 09:50+0000\n"
-"PO-Revision-Date: 2021-07-22 09:50+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: l10n_de_stock
-#: code:addons/l10n_de_stock/models/stock.py:0
-#, python-format
-msgid "Date"
-msgstr ""
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory__id
-msgid "ID"
-msgstr ""
-
-#. module: l10n_de_stock
-#: model:ir.model,name:l10n_de_stock.model_stock_inventory
-msgid "Inventory"
-msgstr ""
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory__l10n_de_template_data
-msgid "L10N De Template Data"
-msgstr ""
-
-#. module: l10n_de_stock
-#: model:ir.model.fields,field_description:l10n_de_stock.field_stock_inventory____last_update
-msgid "Last Modified on"
-msgstr ""
