+++ addons/l10n_din5008/data/report_layout.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <record id="report_layout_din5008" model="report.layout">
+            <field name="view_id" ref="l10n_din5008.external_layout_din5008"/>
+            <field name="image"></field>
+            <field name="pdf"></field>
+            <field name="name">DIN 5008</field>
+        </record>
+    </data>
+</odoo>
