+++ addons/l10n_din5008/static/src/scss/report_din5008.scss
