+++ addons/l10n_din5008_purchase/__init__.py
