+++ addons/l10n_din5008_purchase/__manifest__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+{
+    'name': 'DIN 5008 - Purchase',
+    'category': 'Accounting/Localizations',
+    'depends': [
+        'l10n_din5008',
+        'purchase',
+    ],
+    'auto_install': True,
+    'license': 'LGPL-3',
+}
