+++ addons/l10n_din5008_stock/__init__.py
