+++ addons/l10n_din5008_stock/__manifest__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+{
+    'name': 'DIN 5008 - Stock',
+    'category': 'Accounting/Localizations',
+    'depends': [
+        'l10n_din5008',
+        'stock',
+    ],
+    'auto_install': True,
+    'license': 'LGPL-3',
+}
