+++ addons/l10n_din5008_stock/models/__init__.py
